requirements.txt
python-dotenv
langchain
langchain-openai
langchain_community